refactor: SQL 대신에 명시적인 DslJdbcTemplate 사용
