+++ app/src/main/java/com/techcourse/dao/UserDao.java
@@ -1,8 +1,8 @@
 package com.techcourse.dao;
 
-import com.interface21.jdbc.client.Sql;
 import com.interface21.jdbc.core.JdbcTemplate;
 import com.interface21.jdbc.core.ResultExtractor;
+import com.interface21.jdbc.dsl.DslJdbcTemplate;
 import com.techcourse.domain.User;
 import java.util.List;
 
@@ -15,14 +15,14 @@ public class UserDao {
             rs.getString("email")
     );
 
-    private final Sql sql;
+    private final DslJdbcTemplate jdbc;
 
     public UserDao(final JdbcTemplate jdbcTemplate) {
-        this.sql = new Sql(jdbcTemplate);
+        this.jdbc = new DslJdbcTemplate(jdbcTemplate);
     }
 
     public void insert(final User user) {
-        sql.insert("users")
+        jdbc.insert("users")
                 .of("account", user.getAccount())
                 .of("password", user.getPassword())
                 .of("email", user.getEmail())
@@ -30,7 +30,7 @@ public class UserDao {
     }
 
     public void update(final User user) {
-        sql.update("users")
+        jdbc.update("users")
                 .set("account", user.getAccount())
                 .set("password", user.getPassword())
                 .set("email", user.getEmail())
@@ -39,20 +39,20 @@ public class UserDao {
     }
 
     public List<User> findAll() {
-        return sql.select("id", "account", "password", "email")
+        return jdbc.select("id", "account", "password", "email")
                 .from("users")
                 .many(EXTRACTOR);
     }
 
     public User findById(final Long id) {
-        return sql.select("id", "account", "password", "email")
+        return jdbc.select("id", "account", "password", "email")
                 .from("users")
                 .where("id", id)
                 .one(EXTRACTOR);
     }
 
     public User findByAccount(final String account) {
-        return sql.select("id", "account", "password", "email")
+        return jdbc.select("id", "account", "password", "email")
                 .from("users")
                 .where("account", account)
                 .one(EXTRACTOR);
