+++ jdbc/src/main/java/com/interface21/jdbc/dsl/Select.java
@@ -1,4 +1,4 @@
-package com.interface21.jdbc.client;
+package com.interface21.jdbc.dsl;
 
 import com.interface21.jdbc.core.ResultExtractor;
 import java.util.Arrays;
@@ -8,14 +8,14 @@ import java.util.Map;
 
 public class Select {
 
-    private final Sql sql;
+    private final DslJdbcTemplate dslJdbcTemplate;
 
     private String tableName;
     private final List<String> rowList;
     private final Map<String, Object> whereMap = new HashMap<>();
 
-    public Select(Sql sql, String... rowNames) {
-        this.sql = sql;
+    public Select(DslJdbcTemplate dslJdbcTemplate, String... rowNames) {
+        this.dslJdbcTemplate = dslJdbcTemplate;
         this.rowList = Arrays.stream(rowNames).toList();
     }
 
@@ -31,11 +31,11 @@ public class Select {
 
     // TODO : ResultExtractor를 구현하지 않아도 추출 가능하도록 팩토리 사용
     public <T> List<T> many(ResultExtractor<T> extractor) {
-        return sql.selectMany(this, extractor);
+        return dslJdbcTemplate.selectMany(this, extractor);
     }
 
     public <T> T one(ResultExtractor<T> extractor) {
-        return sql.selectOne(this, extractor);
+        return dslJdbcTemplate.selectOne(this, extractor);
     }
 
     // 완성된 SQL의 형태를 구성한다.
