+++ jdbc/src/main/java/com/interface21/jdbc/dsl/Update.java
@@ -1,18 +1,18 @@
-package com.interface21.jdbc.client;
+package com.interface21.jdbc.dsl;
 
 import java.util.HashMap;
 import java.util.Map;
 
 public class Update {
 
-    private final Sql sql;
+    private final DslJdbcTemplate dslJdbcTemplate;
 
     private final String tableName;
     private final Map<String, Object> setMap = new HashMap<>();
     private final Map<String, Object> whereMap = new HashMap<>();
 
-    public Update(Sql sql, String tableName) {
-        this.sql = sql;
+    public Update(DslJdbcTemplate dslJdbcTemplate, String tableName) {
+        this.dslJdbcTemplate = dslJdbcTemplate;
         this.tableName = tableName;
     }
 
@@ -27,7 +27,7 @@ public class Update {
     }
 
     public void execute() {
-        sql.update(this);
+        dslJdbcTemplate.update(this);
     }
 
     // 완성된 SQL의 형태를 구성한다.
